seperate new lb to a single file
